requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
amqp==5.3.1
    # via kombu
asgiref==3.8.1
    # via django
asttokens==3.0.0
    # via stack-data
babel==2.16.0
    # via py-moneyed
billiard==4.2.1
    # via celery
black==24.10.0
    # via -r requirements.in
build==1.2.2.post1
    # via pip-tools
celery==5.4.0
    # via
    #   django-celery-beat
    #   flower
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   nltk
    #   pip-tools
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
colorama==0.4.6
    # via
    #   build
    #   click
    #   ipython
    #   tqdm
crispy-bootstrap5==2024.10
    # via -r requirements.in
cron-descriptor==1.4.5
    # via django-celery-beat
cryptography==44.0.0
    # via social-auth-core
decorator==5.1.1
    # via ipython
defusedxml==0.8.0rc2
    # via
    #   python3-openid
    #   social-auth-core
django==4.2.16
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-celery-beat
    #   django-crispy-forms
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-money
    #   django-phonenumber-field
    #   django-timezone-field
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   djoser
    #   drf-yasg
    #   social-auth-app-django
django-admin-interface==0.29.2
    # via -r requirements.in
django-celery-beat==2.7.0
    # via -r requirements.in
django-colorfield==0.11.0
    # via django-admin-interface
django-crispy-forms==2.3
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-debug-toolbar==4.4.6
    # via -r requirements.in
django-extensions==3.2.3
    # via -r requirements.in
django-filter==24.3
    # via -r requirements.in
django-money==3.5.3
    # via -r requirements.in
django-phonenumber-field==8.0.0
    # via -r requirements.in
django-templated-mail==1.1.1
    # via -r requirements.in
django-timezone-field==7.0
    # via django-celery-beat
djangorestframework==3.15.2
    # via
    #   djangorestframework-simplejwt
    #   drf-yasg
djangorestframework-simplejwt==5.3.1
    # via djoser
djoser==2.3.1
    # via -r requirements.in
dnspython==2.7.0
    # via pymongo
drf-yasg==1.21.8
    # via -r requirements.in
executing==2.1.0
    # via stack-data
faker==33.1.0
    # via -r requirements.in
flake8==7.1.1
    # via -r requirements.in
flower==2.0.1
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
humanize==4.11.0
    # via flower
idna==3.10
    # via requests
inflection==0.5.1
    # via drf-yasg
ipython==8.30.0
    # via -r requirements.in
isort==5.13.2
    # via -r requirements.in
jedi==0.19.2
    # via ipython
joblib==1.4.2
    # via nltk
kombu==5.4.2
    # via celery
markdown==3.7
    # via -r requirements.in
matplotlib-inline==0.1.7
    # via ipython
mccabe==0.7.0
    # via flake8
mongoengine==0.29.1
    # via -r requirements.in
mypy-extensions==1.0.0
    # via black
nltk==3.9.1
    # via -r requirements.in
oauthlib==3.2.2
    # via
    #   requests-oauthlib
    #   social-auth-core
packaging==24.2
    # via
    #   black
    #   build
    #   drf-yasg
    #   gunicorn
parso==0.8.4
    # via jedi
pathspec==0.12.1
    # via black
phonenumbers==8.13.51
    # via -r requirements.in
pillow==11.0.0
    # via django-colorfield
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.3.6
    # via black
prometheus-client==0.21.0
    # via flower
prompt-toolkit==3.0.48
    # via
    #   click-repl
    #   ipython
psycopg2-binary==2.9.10
    # via -r requirements.in
pure-eval==0.2.3
    # via stack-data
py-moneyed==3.0
    # via django-money
pycodestyle==2.12.1
    # via flake8
pycparser==2.22
    # via cffi
pydot==3.0.3
    # via -r requirements.in
pyflakes==3.2.0
    # via flake8
pygments==2.18.0
    # via ipython
pyjwt==2.10.1
    # via
    #   djangorestframework-simplejwt
    #   social-auth-core
pymongo==4.10.1
    # via mongoengine
pyparsing==3.2.0
    # via pydot
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-crontab==3.2.0
    # via django-celery-beat
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   faker
    #   python-crontab
python-slugify==8.0.4
    # via django-admin-interface
python3-openid==3.2.0
    # via social-auth-core
pytz==2024.2
    # via
    #   drf-yasg
    #   flower
pyyaml==6.0.2
    # via drf-yasg
redis==5.2.0
    # via -r requirements.in
regex==2024.11.6
    # via nltk
requests==2.32.3
    # via
    #   requests-oauthlib
    #   social-auth-core
requests-oauthlib==2.0.0
    # via social-auth-core
six==1.16.0
    # via python-dateutil
social-auth-app-django==5.4.2
    # via djoser
social-auth-core==4.5.4
    # via social-auth-app-django
sqlparse==0.5.2
    # via
    #   django
    #   django-debug-toolbar
stack-data==0.6.3
    # via ipython
text-unidecode==1.3
    # via python-slugify
tornado==6.4.2
    # via flower
tqdm==4.67.1
    # via nltk
traitlets==5.14.3
    # via
    #   ipython
    #   matplotlib-inline
typing-extensions==4.12.2
    # via
    #   faker
    #   py-moneyed
tzdata==2024.2
    # via
    #   celery
    #   django
    #   django-celery-beat
    #   kombu
uritemplate==4.1.1
    # via drf-yasg
urllib3==2.2.3
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
